+++ Src/Private/Expand-LabImage.ps1
@@ -68,6 +68,9 @@ function Expand-LabImage {
 
             if ($mediaFileInfo.Extension -eq '.ISO') {
 
+                ## Disable BitLocker fixed drive write protection (if enabled)
+                Disable-BitLockerFDV;
+
                 ## Mount ISO
                 Write-Verbose -Message ($localized.MountingDiskImage -f $MediaPath);
                 $mountDiskImageParams = @{
@@ -114,8 +117,7 @@ function Expand-LabImage {
                 Verbose = $false;
                 ErrorAction = 'Stop';
             }
-            $dismOutput = Expand-WindowsImage @expandWindowsImageParams;
-
+            [ref] $null = Expand-WindowsImage @expandWindowsImageParams;
             [ref] $null = Get-PSDrive;
 
             ## Add additional packages (.cab) files
@@ -135,7 +137,7 @@ function Expand-LabImage {
                     ## Use the specified/literal path
                     $addLabImageWindowsPackageParams['PackagePath'] = $PackagePath;
                 }
-                $dismOutput = Add-LabImageWindowsPackage @addLabImageWindowsPackageParams;
+                [ref] $null = Add-LabImageWindowsPackage @addLabImageWindowsPackageParams;
 
             } #end if Package
 
@@ -155,7 +157,7 @@ function Expand-LabImage {
                     ## The Windows optional feature source path for .WIM files is a literal path
                     $addLabImageWindowsOptionalFeatureParams['ImagePath'] = $SourcePath;
                 }
-                $dismOutput = Add-LabImageWindowsOptionalFeature @addLabImageWindowsOptionalFeatureParams;
+                [ref] $null = Add-LabImageWindowsOptionalFeature @addLabImageWindowsOptionalFeatureParams;
             } #end if WindowsOptionalFeature
 
         }
@@ -172,6 +174,9 @@ function Expand-LabImage {
                 Write-Verbose -Message ($localized.DismountingDiskImage -f $MediaPath);
                 Storage\Dismount-DiskImage -ImagePath $MediaPath;
             }
+
+            ## Enable BitLocker (if required)
+            Assert-BitLockerFDV
 
         } #end finally