+++ Src/Private/New-DiskImageGpt.ps1
@@ -4,6 +4,7 @@ function New-DiskImageGpt {
         Create a new GPT-formatted disk image.
 #>
     [CmdletBinding()]
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions','')]
     param (
         ## Mounted VHD(X) Operating System disk image
         [Parameter(Mandatory)]