+++ Src/Private/Set-DiskImageBootVolumeGpt.ps1
@@ -4,6 +4,7 @@ function Set-DiskImageBootVolumeGpt {
         Configure/repair MBR boot volume
 #>
     [CmdletBinding()]
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions','')]
     param (
         ## Mounted VHD(X) Operating System disk image
         [Parameter(Mandatory)]