+++ Tests/Unit/Src/Private/Expand-LabIso.Tests.ps1
@@ -18,6 +18,7 @@ Describe 'Src\Private\Expand-LabIso' {
             Mock CopyDirectory -MockWith { }
             Mock Dismount-DiskImage -MockWith { }
             Mock Mount-DiskImage -ParameterFilter { $ImagePath -eq $testIsoPath -and $Access -eq 'ReadOnly' } -MockWith { return $fakeDiskImage; }
+            Mock Get-ItemProperty -MockWith { [PSCustomObject] @{ FDVDenyWriteAccess = 0 } } -ParameterFilter {$Name -eq 'FDVDenyWriteAccess'}
 
             Expand-LabIso -Path $testIsoPath -DestinationPath $testDestinationPath;
 
@@ -33,6 +34,7 @@ Describe 'Src\Private\Expand-LabIso' {
             Mock Get-Volume -MockWith { return [PSCustomObject] @{ DriveLetter = $testIsoMountDrive } }
             Mock Dismount-DiskImage -MockWith { }
             Mock CopyDirectory -ParameterFilter { $DestinationPath.FullName -eq $testDestinationPath } -MockWith { }
+            Mock Get-ItemProperty -MockWith { [PSCustomObject] @{ FDVDenyWriteAccess = 0 } } -ParameterFilter {$Name -eq 'FDVDenyWriteAccess'}
 
             Expand-LabIso -Path $testIsoPath -DestinationPath $testDestinationPath;
 
@@ -48,10 +50,29 @@ Describe 'Src\Private\Expand-LabIso' {
             Mock Get-Volume -MockWith { return [PSCustomObject] @{ DriveLetter = $testIsoMountDrive } }
             Mock CopyDirectory -MockWith { }
             Mock Dismount-DiskImage -ParameterFilter { $ImagePath -eq $testIsoPath } -MockWith { }
+            Mock Get-ItemProperty -MockWith { [PSCustomObject] @{ FDVDenyWriteAccess = 0 } } -ParameterFilter {$Name -eq 'FDVDenyWriteAccess'}
 
             Expand-LabIso -Path $testIsoPath -DestinationPath $testDestinationPath;
 
             Assert-MockCalled Dismount-DiskImage -ParameterFilter { $ImagePath -eq $testIsoPath } -Scope It;
+        }
+
+        It 'Disables/enables BitLocker fixed drive write protection (if required)' {
+            $testIsoPath = 'TestDrive:\TestIsoImage.iso';
+            $testDestinationPath = 'Y:\';
+            $testIsoMountDrive = 'Z';
+            $fakeDiskImage = [PSCustomObject] @{ DriveLetter = $testIsoMountDrive; ImagePath = $testIsoPath };
+            Mock Get-Volume -MockWith { return [PSCustomObject] @{ DriveLetter = $testIsoMountDrive } }
+            Mock CopyDirectory -MockWith { }
+            Mock Dismount-DiskImage -MockWith { }
+            Mock Mount-DiskImage -ParameterFilter { $ImagePath -eq $testIsoPath -and $Access -eq 'ReadOnly' } -MockWith { return $fakeDiskImage; }
+            Mock Disable-BitLockerFDV -MockWith { }
+            Mock Assert-BitLockerFDV -MockWith { }
+
+            Expand-LabIso -Path $testIsoPath -DestinationPath $testDestinationPath;
+
+            Assert-MockCalled Disable-BitLockerFDV -Exactly 1 -Scope It
+            Assert-MockCalled Assert-BitLockerFDV -Exactly 1 -Scope It
         }
 
     } #end InModuleScope